Update 404.html
redirecting to the root is enough for now because while a User might want to share a link like "foxcycle.github.io/productview/3" with a friend... this is hard with GitHub Pages which are meant to be static html sites rather than dynamic Vue sites. To fulfill that use case, a real server should be used instead of gh-pages.