simply redirect to the client server
so it turns out that there is an error in the console when attempting to perform a GET request to http://<API_IP_ADDRESS>/<route> because github only allows https since the url is https://foxcycle.github.io

so the possible solutions are to:

1. make an official ssl certificate through an authority and then set that up within docker for the mysql server.

2. do a "self-signed" ssl cert and set that up within the docker container for the mysql server

3. ditch docker entirely and set up the mysql server without docker and then also set up the ssl cert

4. forget about trying to make github pages work with the Foxcycle app and simply redirect to the IP address.

......

I'm going with option 4 because I read about how to do the other options and I am stuck on how to set up ssl within a docker container (if that is even a reasonable thing to attempt). Also, I would rather not have to configure the mysql server without docker.